formatting to render correctly on the server and change a few links around for better organization
