+++ app/templates/navbar.html
@@ -38,18 +38,20 @@
     </li>
     <li class="nav-item dropdown">
       <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Support</a>
-      <div class="dropdown-menu dropdown-menu-right navbar-dark bg-dark" aria-labelledby="navbarDropdown">
-        <a class="dropdown-item nav-link bg-dark" href="/support_accessories">Accessories</a>
-        <a class="dropdown-item nav-link bg-dark" href="/support_additional">Additional Resources</a>
-        <a class="dropdown-item nav-link bg-dark" href="/support_kegsmarts">KegSmarts</a>
-        <a class="dropdown-item nav-link bg-dark" href="/support_pico_c">Pico C</a>
-        <a class="dropdown-item nav-link bg-dark" href="/support_pico_pro">Pico Pro</a>
-        <a class="dropdown-item nav-link bg-dark" href="/support_pico_s">Pico S</a>
-        <a class="dropdown-item nav-link bg-dark" href="/support_picoferm">PicoFerm</a>
-        <a class="dropdown-item nav-link bg-dark" href="/support_picostill">PicoStill</a>
-        <a class="dropdown-item nav-link bg-dark" href="/support_z_series">Z Series</a>
-        <a class="dropdown-item nav-link bg-dark" href="/support_zymatic">Zymatic</a>
-      </div>
+      <ul class="dropdown-menu dropdown-menu-right navbar-dark bg-dark" aria-labelledby="navbarDropdown">
+        <li><a class="dropdown-item nav-link bg-dark text-white" href="/support/accessories">Accessories</a></li>
+        <li><a class="dropdown-item nav-link bg-dark text-white" href="/support/additional_info">Additional Resources</a></li>
+        <li role="separator" class="drodown-divider"></li>
+        <!-- Unsupported device, though support docs are loaded into server
+        <li><a class="dropdown-item nav-link bg-dark text-white" href="/support_kegsmarts">KegSmarts</a></li> -->
+        <li><a class="dropdown-item nav-link bg-dark text-white" href="/support/pico_c">Pico C</a></li>
+        <li><a class="dropdown-item nav-link bg-dark text-white" href="/support/pico_pro">Pico Pro</a></li>
+        <li><a class="dropdown-item nav-link bg-dark text-white" href="/support/pico_s">Pico S</a></li>
+        <li><a class="dropdown-item nav-link bg-dark text-white" href="/support/picoferm">PicoFerm</a></li>
+        <li><a class="dropdown-item nav-link bg-dark text-white" href="/support/picostill">PicoStill</a></li>
+        <li><a class="dropdown-item nav-link bg-dark text-white" href="/support/z_series">Z Series</a></li>
+        <li><a class="dropdown-item nav-link bg-dark text-white" href="/support/zymatic">Zymatic</a></li>
+      </ul>
     </li>
     <li class="nav-item dropdown">
       <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">System</a>
